pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
requires-python = ">=3.8"
dependencies = [
  "pillow>=12.0.0",
]

